api_yamdb/api/filters.py
from django_filters import rest_framework as filters
from reviews.models import Title


class FilterForTitle(filters.FilterSet):
    """Custom filter for fields"""
    category = filters.CharFilter(field_name='category__slug')
    genre = filters.CharFilter(field_name='genre__slug')
    name = filters.CharFilter(field_name='name', lookup_expr='contains')
    year = filters.NumberFilter(field_name='year')

    class Meta:
        model = Title
        fields = ('category', 'genre', 'name', 'year')
